pyproject.toml
[project]
name = "spotify-wrapped"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "spotipy>=2.24.0",
    "streamlit>=1.35.0",
    "pandas>=2.2.0",
    "altair>=5.3.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wrapped"]
